Replace sample logger with firebase crashlytics logger
