+++ stream-chat-android-sample/src/main/kotlin/io/getstream/chat/sample/application/ChatInitializer.kt
@@ -18,7 +18,7 @@ class ChatInitializer(private val context: Context) {
                     smallIcon = R.drawable.ic_chat_bubble
                 )
             notificationHandler = SampleNotificationHandler(context, notificationConfig)
-            chatLoggerHandler = SampleLoggingHandler()
+            chatLoggerHandler = FirebaseLogger
         }.build()
     }
 }
